Try detecting libffi with pkg-config first, fall-back to plainly use the library
In some GNU/Linux distributions (example: Arch Linux) the headers for
libffi are installed in a location which is not in the default compiler
search path, and a libffi.pc file is provided to allow discovery of the
needed build flags using pkg-config.

With this fix applied, Meson will try first to find libffi using
pkg-config by means of dependency(), and if not found, fall-back to
compiler.find_library() which will still cover systems which do not
install a .pc file and provide the libffi headers in a standard
location.

This fixes the following compilation error:

  FAILED: xsdump@exe/src_prim-etc.cxx.o
  c++  -Ixsdump@exe -I. -I.. -I../src -fdiagnostics-color=always -pipe -D_FILE_OFFSET_BITS=64 -Wall -Winvalid-pch -Wnon-virtual-dtor -O0 -g -Wall -Wextra -MMD -MQ 'xsdump@exe/src_prim-etc.cxx.o' -MF 'xsdump@exe/src_prim-etc.cxx.o.d' -o 'xsdump@exe/src_prim-etc.cxx.o' -c ../src/prim-etc.cxx
../src/prim-etc.cxx:8:10: fatal error: ffi.h: No such file or directory
   #include <ffi.h>
            ^~~~~~~